+++ components/sql/src/test/scala/pl/touk/nussknacker/sql/utils/StandaloneProcessTest.scala
@@ -2,16 +2,19 @@ package pl.touk.nussknacker.sql.utils
 
 import org.scalatest.Matchers
 import org.scalatest.concurrent.ScalaFutures
+import pl.touk.nussknacker.engine.api.ProcessVersion
 import pl.touk.nussknacker.engine.api.deployment.DeploymentData
 import pl.touk.nussknacker.engine.api.process.RunMode
-import pl.touk.nussknacker.engine.api.{JobData, ProcessVersion}
 import pl.touk.nussknacker.engine.baseengine.api.runtimecontext.EngineRuntimeContextPreparer
 import pl.touk.nussknacker.engine.graph.EspProcess
 import pl.touk.nussknacker.engine.resultcollector.ProductionServiceInvocationCollector
+import pl.touk.nussknacker.engine.standalone.FutureBaseStandaloneScenarioEngine._
 import pl.touk.nussknacker.engine.standalone.StandaloneScenarioEngine
 import pl.touk.nussknacker.engine.standalone.StandaloneScenarioEngine.StandaloneResultType
 import pl.touk.nussknacker.engine.testing.LocalModelData
 import pl.touk.nussknacker.engine.util.SynchronousExecutionContext.ctx
+
+import scala.concurrent.Future
 
 trait StandaloneProcessTest extends Matchers with ScalaFutures {
 
@@ -31,8 +34,8 @@ trait StandaloneProcessTest extends Matchers with ScalaFutures {
     }
   }
 
-  private def prepareInterpreter(process: EspProcess): StandaloneScenarioEngine.StandaloneScenarioInterpreter = {
-    val validatedInterpreter = StandaloneScenarioEngine(process,
+  private def prepareInterpreter(process: EspProcess): InterpreterType = {
+    val validatedInterpreter = StandaloneScenarioEngine[Future](process,
       ProcessVersion.empty, DeploymentData.empty,
       contextPreparer, modelData, Nil, ProductionServiceInvocationCollector, runMode)