+++ engine/base/runtime/src/main/scala/pl/touk/nussknacker/engine/baseengine/TestRunner.scala
@@ -1,5 +1,6 @@
 package pl.touk.nussknacker.engine.baseengine
 
+import cats.{Id, ~>}
 import cats.data.Validated.{Invalid, Valid}
 import pl.touk.nussknacker.engine.Interpreter.InterpreterShape
 import pl.touk.nussknacker.engine.ModelData
@@ -7,6 +8,7 @@ import pl.touk.nussknacker.engine.api.deployment.DeploymentData
 import pl.touk.nussknacker.engine.api.deployment.TestProcess.{TestData, TestResults}
 import pl.touk.nussknacker.engine.api.process.{ProcessName, RunMode, Source}
 import pl.touk.nussknacker.engine.api.{JobData, ProcessVersion}
+import pl.touk.nussknacker.engine.baseengine.TestRunner.EffectUnwrapper
 import pl.touk.nussknacker.engine.baseengine.api.commonTypes.ResultType
 import pl.touk.nussknacker.engine.baseengine.api.customComponentTypes.CapabilityTransformer
 import pl.touk.nussknacker.engine.baseengine.api.interpreterTypes.{EndResult, ScenarioInputBatch, SourceId}
@@ -20,11 +22,9 @@ import scala.language.higherKinds
 
 //TODO: integrate with Engine somehow?
 //Base test runner, creating Context from sampleData and mapping results are left for the implementations for now
-abstract class TestRunner[F[_], Res <: AnyRef](shape: InterpreterShape[F], capabilityTransformer: CapabilityTransformer[F]) {
+abstract class TestRunner[F[_] : InterpreterShape : CapabilityTransformer : EffectUnwrapper, Res <: AnyRef] {
 
   def sampleToSource(sampleData: List[AnyRef], sources: Map[SourceId, Source]): ScenarioInputBatch
-
-  def getResults(results: F[ResultType[EndResult[Res]]]): ResultType[EndResult[Res]]
 
   def runTest[T](modelData: ModelData,
                  testData: TestData,
@@ -42,7 +42,7 @@ abstract class TestRunner[F[_], Res <: AnyRef](shape: InterpreterShape[F], capab
     //FIXME: validation??
     val standaloneInterpreter = ScenarioInterpreterFactory.createInterpreter[F, Res](process, modelData,
       additionalListeners = List(collectingListener), new TestServiceInvocationCollector(collectingListener.runId), runMode
-    )(SynchronousExecutionContext.ctx, shape, capabilityTransformer) match {
+    )(SynchronousExecutionContext.ctx, implicitly[InterpreterShape[F]], implicitly[CapabilityTransformer[F]]) match {
       case Valid(interpreter) => interpreter
       case Invalid(errors) => throw new IllegalArgumentException("Error during interpreter preparation: " + errors.toList.mkString(", "))
     }
@@ -52,7 +52,7 @@ abstract class TestRunner[F[_], Res <: AnyRef](shape: InterpreterShape[F], capab
 
       val inputs = sampleToSource(parsedTestData.samples, standaloneInterpreter.sources)
 
-      val results = getResults(standaloneInterpreter.invoke(inputs))
+      val results = implicitly[EffectUnwrapper[F]].apply(standaloneInterpreter.invoke(inputs))
 
       collectSinkResults(collectingListener.runId, results)
       collectingListener.results
@@ -78,5 +78,11 @@ abstract class TestRunner[F[_], Res <: AnyRef](shape: InterpreterShape[F], capab
       SinkInvocationCollector(runId, node, node).collect(result.context, result.result)
     }
   }
+
+}
+
+object TestRunner {
+
+  type EffectUnwrapper[F[_]] = F ~> Id
 
 }