+++ engine/standalone/app/src/main/scala/pl/touk/nussknacker/engine/standalone/http/StandaloneRequestHandler.scala
@@ -5,8 +5,10 @@ import cats.data.{NonEmptyList, Validated}
 import io.circe.Json
 import pl.touk.nussknacker.engine.api.Context
 import pl.touk.nussknacker.engine.api.exception.EspExceptionInfo
+import pl.touk.nussknacker.engine.standalone.FutureBaseStandaloneScenarioEngine.InterpreterType
 import pl.touk.nussknacker.engine.standalone.StandaloneScenarioEngine.StandaloneResultType
 import pl.touk.nussknacker.engine.standalone.api.{StandaloneGetSource, StandalonePostSource}
+import pl.touk.nussknacker.engine.standalone.metrics.InvocationMetrics
 import pl.touk.nussknacker.engine.standalone.{DefaultResponseEncoder, StandaloneScenarioEngine}
 
 import scala.concurrent.{ExecutionContext, Future}
@@ -15,11 +17,13 @@ import scala.util.Try
 
 //this class handles parsing, displaying and invoking interpreter. This is the only place we interact with model, hence
 //only here we care about context classloaders
-class StandaloneRequestHandler(standaloneProcessInterpreter: StandaloneScenarioEngine.StandaloneScenarioInterpreter) extends Directives  {
+class StandaloneRequestHandler(standaloneProcessInterpreter: InterpreterType) extends Directives  {
 
   private val source = standaloneProcessInterpreter.source
 
   private val encoder = source.responseEncoder.getOrElse(DefaultResponseEncoder)
+
+  private val invocationMetrics = new InvocationMetrics(standaloneProcessInterpreter.context)
 
   private val extractInput: Directive1[Any] = source match {
     case a: StandalonePostSource[Any] =>
@@ -35,9 +39,10 @@ class StandaloneRequestHandler(standaloneProcessInterpreter: StandaloneScenarioE
         .flatMap(onSuccess(_))
     }
 
-  private def invokeInterpreter(input: Any)(implicit ec: ExecutionContext): Future[StandaloneScenarioEngine.StandaloneResultType[Json]] =
+  private def invokeInterpreter(input: Any)(implicit ec: ExecutionContext): Future[StandaloneScenarioEngine.StandaloneResultType[Json]] = invocationMetrics.measureTime {
     standaloneProcessInterpreter.invokeToOutput(input).map(_.andThen { data =>
       Validated.fromTry(Try(encoder.toJsonResponse(input, data))).leftMap(ex => NonEmptyList.one(EspExceptionInfo(None, ex, Context(""))))
     })
+  }
 
 }